move spec files to subfolders
