+++ e2e/specs/wallet/request-token-flow.spec.js
@@ -1,18 +1,18 @@
 'use strict';
-import { Smoke } from '../tags';
+import { Smoke } from '../../tags';
 
-import SendLinkView from '../pages/SendLinkView';
-import RequestPaymentView from '../pages/RequestPaymentView';
+import SendLinkView from '../../pages/SendLinkView';
+import RequestPaymentView from '../../pages/RequestPaymentView';
 
-import TabBarComponent from '../pages/TabBarComponent';
-import WalletActionsModal from '../pages/modals/WalletActionsModal';
+import TabBarComponent from '../../pages/TabBarComponent';
+import WalletActionsModal from '../../pages/modals/WalletActionsModal';
 
-import ProtectYourWalletModal from '../pages/modals/ProtectYourWalletModal';
-import RequestPaymentModal from '../pages/modals/RequestPaymentModal';
+import ProtectYourWalletModal from '../../pages/modals/ProtectYourWalletModal';
+import RequestPaymentModal from '../../pages/modals/RequestPaymentModal';
 
-import { loginToApp } from '../viewHelper';
-import { withFixtures } from '../fixtures/fixture-helper';
-import FixtureBuilder from '../fixtures/fixture-builder';
+import { loginToApp } from '../../viewHelper';
+import { withFixtures } from '../../fixtures/fixture-helper';
+import FixtureBuilder from '../../fixtures/fixture-builder';
 
 const SAI_CONTRACT_ADDRESS = '0x89d24a6b4ccb1b6faa2625fe562bdd9a23260359';
