+++ e2e/specs/wallet/send-ERC-token.spec.js
@@ -1,17 +1,17 @@
 'use strict';
-import { Smoke } from '../tags';
+import { Smoke } from '../../tags';
 
-import TestHelpers from '../helpers';
+import TestHelpers from '../../helpers';
 
-import WalletView from '../pages/WalletView';
-import NetworkEducationModal from '../pages/modals/NetworkEducationModal';
-import AddCustomTokenView from '../pages/AddCustomTokenView';
-import SendView from '../pages/SendView';
-import AmountView from '../pages/AmountView';
-import { importWalletWithRecoveryPhrase } from '../viewHelper';
-import TransactionConfirmationView from '../pages/TransactionConfirmView';
-import NetworkListModal from '../pages/modals/NetworkListModal';
-import TokenOverview from '../pages/TokenOverview';
+import WalletView from '../../pages/WalletView';
+import NetworkEducationModal from '../../pages/modals/NetworkEducationModal';
+import AddCustomTokenView from '../../pages/AddCustomTokenView';
+import SendView from '../../pages/SendView';
+import AmountView from '../../pages/AmountView';
+import { importWalletWithRecoveryPhrase } from '../../viewHelper';
+import TransactionConfirmationView from '../../pages/TransactionConfirmView';
+import NetworkListModal from '../../pages/modals/NetworkListModal';
+import TokenOverview from '../../pages/TokenOverview';
 
 const TOKEN_ADDRESS = '0x779877A7B0D9E8603169DdbD7836e478b4624789';
 const SEND_ADDRESS = '0xebe6CcB6B55e1d094d9c58980Bc10Fed69932cAb';
