+++ e2e/specs/wallet/start-exploring.spec.js
@@ -1,23 +1,23 @@
 'use strict';
-import { Smoke } from '../tags';
+import { Smoke } from '../../tags';
 
-import TestHelpers from '../helpers';
+import TestHelpers from '../../helpers';
 
-import OnboardingView from '../pages/Onboarding/OnboardingView';
-import OnboardingCarouselView from '../pages/Onboarding/OnboardingCarouselView';
-import ProtectYourWalletView from '../pages/Onboarding/ProtectYourWalletView';
-import CreatePasswordView from '../pages/Onboarding/CreatePasswordView';
+import OnboardingView from '../../pages/Onboarding/OnboardingView';
+import OnboardingCarouselView from '../../pages/Onboarding/OnboardingCarouselView';
+import ProtectYourWalletView from '../../pages/Onboarding/ProtectYourWalletView';
+import CreatePasswordView from '../../pages/Onboarding/CreatePasswordView';
 
-import MetaMetricsOptIn from '../pages/Onboarding/MetaMetricsOptInView';
-import WalletView from '../pages/WalletView';
-import EnableAutomaticSecurityChecksView from '../pages/EnableAutomaticSecurityChecksView';
+import MetaMetricsOptIn from '../../pages/Onboarding/MetaMetricsOptInView';
+import WalletView from '../../pages/WalletView';
+import EnableAutomaticSecurityChecksView from '../../pages/EnableAutomaticSecurityChecksView';
 
-import Browser from '../pages/Drawer/Browser';
+import Browser from '../../pages/Drawer/Browser';
 
-import SkipAccountSecurityModal from '../pages/modals/SkipAccountSecurityModal';
-import OnboardingWizardModal from '../pages/modals/OnboardingWizardModal';
-import WhatsNewModal from '../pages/modals/WhatsNewModal';
-import { acceptTermOfUse } from '../viewHelper';
+import SkipAccountSecurityModal from '../../pages/modals/SkipAccountSecurityModal';
+import OnboardingWizardModal from '../../pages/modals/OnboardingWizardModal';
+import WhatsNewModal from '../../pages/modals/WhatsNewModal';
+import { acceptTermOfUse } from '../../viewHelper';
 
 const ACCOUNT = 'Test Account One';
 const PASSWORD = '12345678';
